agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
google-generativeai==0.7.2
python-dotenv==1.0.1
